06/15/2015
This code is based on modification of the code of 05/07/2015

1. \hat{V} is now estimated using repeated generated samples instead of
bootstrap.

2. Rao's 1st and 2nd correction is now calculated using the estimates of
eigenvalues instead of using eigen() function in R directly. The details
of how to calculate these estimates can be found in report of 20150615.

3. For 2nd order critical value, df is not an integer, but same results
will be given using qchisq() or qgamma() functions.

4. The parameter T is now a vector which contains 3 number of iterations
for total iteration, iteration of W and number of repeated samples for
calculating \hat{V}. For example, T=c(10000, 100000, 10000) means that
total loop is 10,000, critical values for W is calculated using 100,000
iterations, and \hat{V} is estimated using 10,000 repeated samples.