09/08/2015
This code is based on modification of the code of 09/01/2015

1. It is comfirmed that R eigen() function is equivalent to the manual
method by Rao.

2. Code for simulating weighted a_alpha by both 1992 and 1997 of Eubank
is developed in a_table.R. But neither of them gave the correct values.
I guess this is because weighted a_alpha may not be from these 2
equations.

3. I proposed another easier method, which can find a_alpha correctly.
It is Weighted a_alpha = \delta_{.} x Unweighted a_alpha.

4. The code for simulating the weighted version of those 3 alternatives
from Euabnk (1997) are developed in eubank.run, which was re-organized
for easier reading and excution.