09/15/2015
This code is based on modification of the code of 09/08/2015

1. A bug is fixed for the equation of Euabnk (1997). Instead of using
"rchisq(1, df=k)", I used the "sum of rchisq(1, df=1)". Now, it is
comfirmed that the equation of Eubank (1997) can also generate
a_{0.05}=4.18.

2. The Unweighted a_alpha is now correctly produced by the equation from
Euabnk (1997). The code is in a_table.R, which contains 2 options in the
loop. I suggest to use the method delta_dot*chisq1.

3. A new function "get_eigen()" is created to calcualte delta_dot,
asquare and all eigenvalues for given data.

4. The Weighted a_alpha is now correctly produced by the modification of
Eubank (1997). The code now utilizes the new created "get_eigen()" to
calculate all eigenvalues, delta_dot, and asquare.

5. The newly created function "get_eigen()" is now embedded into the
functions "eubank()" and "rao()" to replace the chunk of code which
generates detal_dot, asquare and all eigenvalues.

6. In eubank_run.R, both Unweighted and Weighted cases are merged
together. To change between them, just change the value of rho, that is
, if rho=0, it is Unweighted, but if rho>0, it is Weighted.